refactor(layout,#171): Phase B - generalize manifest validation rule & relocate LayoutEngine implementation
Changes:
- ESLint require-manifest-validation now supports domain path and exempts pure re-export shims
- Added tests for legacy path, domain path, and shim exemption
- Moved LayoutEngine implementation to src/domain/layout/LayoutEngine.tsx
- Converted root src/layout/LayoutEngine.tsx into transitional re-export shim

All tests and build green (169 passed, 1 skipped).
